cloud build: set up Go with helper code from prow.sh
This introduces .cloudbuild.sh as the main entry point for building
cloud images, similar to the role that .prow.sh has for Prow testing.

The script sets up the build environment (gcloud auth, installing Go,
parsing some of the env variables) before invoking "make
push-multiarch", which (as before) is usable also outside of a cloud
build.