add --timeout for gRPC calls
The one second timeout is based on the assumption that once the
connection is established, all further communication is going to be
fast.

That assumption doesn't hold in some special circumstances:
- interactively debugging the sidecar
- proxying the driver operation to some remote implementation

While this is a bit obscure, the flag is essential for these
cases. Adding the flag also makes the node-driver-registrar consistent
with other sidecars.